fix text nodes exports
