+++ packages/plugins/lists/src/plugin/TodoList.tsx
@@ -1,4 +1,11 @@
-import { buildBlockData, generateId, YooptaBlockData, YooptaPlugin } from '@yoopta/editor';
+import {
+  buildBlockData,
+  deserializeTextNodes,
+  generateId,
+  serializeTextNodes,
+  YooptaBlockData,
+  YooptaPlugin,
+} from '@yoopta/editor';
 import { TodoListCommands } from '../commands';
 import { TodoListRender } from '../elements/TodoList';
 import { onKeyDown } from '../events/onKeyDown';
@@ -28,10 +35,8 @@ const TodoList = new YooptaPlugin<Pick<ListElementMap, 'todo-list'>>({
   parsers: {
     html: {
       deserialize: {
-        // add ignore or continue statement
         nodeNames: ['OL', 'UL'],
-        // nodeNames: [],
-        parse(el) {
+        parse(el, editor) {
           if (el.nodeName === 'OL' || el.nodeName === 'UL') {
             const listItems = el.querySelectorAll('li');
 
@@ -48,7 +53,6 @@ const TodoList = new YooptaPlugin<Pick<ListElementMap, 'todo-list'>>({
               .map((listItem) => {
                 const textContent = listItem.textContent || '';
                 const checked = /\[\s*x\s*\]/i.test(textContent);
-                const clearedContent = textContent.replace(/\[\s*\S?\s*\]/, '').trim();
 
                 return buildBlockData({
                   id: generateId(),
@@ -57,7 +61,7 @@ const TodoList = new YooptaPlugin<Pick<ListElementMap, 'todo-list'>>({
                     {
                       id: generateId(),
                       type: 'todo-list',
-                      children: [{ text: clearedContent }],
+                      children: deserializeTextNodes(editor, listItem.childNodes),
                       props: { nodeType: 'block', checked: checked },
                     },
                   ],
@@ -74,7 +78,7 @@ const TodoList = new YooptaPlugin<Pick<ListElementMap, 'todo-list'>>({
 
         return `<ul data-meta-align="${align}" data-meta-depth="${depth}" style="margin-left: ${depth}px; text-align: ${align}"><li>[${
           element.props.checked ? 'x' : ' '
-        }] ${text}</li></ul>`;
+        }] ${serializeTextNodes(element.children)}</li></ul>`;
       },
     },
     markdown: {
